Next round of battery testing 12/12
Scale reading to perform 5 readings using an average of 3 measurements.
This reduces the number of scale readings from 20 to 15.
Only read the scale every 16 seconds instead of 8.
Only send confirmation every 4 hours instead of every hour.
Don't power the phone off in startup() as this seems to de-stabilise it
when we immediately power it up again to send the first reading,
particularly when DEBUG was not enabled.
Simplified scale zeroing logic to be two consecutive negative readings
or one big negative reading.
If send to remote server times out then don't clear the readings so that
they get sent next time.
Changed larger delays to use goToSleep to try and save power.
Changed all phone power off calls to use a hardware power off as it
seems more reliable.
Put conditional DEBUG around all log functions so they don't log to
serial if it's not DEBUG mode to try and save power.